refactor: 도메인별 그룹 내용에 맞게 패키지 수정
