+++ src/main/java/kitchenpos/menu/application/ProductService.java
@@ -1,8 +1,8 @@
-package kitchenpos.product.application;
+package kitchenpos.menu.application;
 
-import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.product.dto.ProductRequest;
-import kitchenpos.product.dto.ProductResponse;
+import kitchenpos.menu.domain.ProductRepository;
+import kitchenpos.menu.dto.ProductRequest;
+import kitchenpos.menu.dto.ProductResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
