+++ src/main/java/kitchenpos/menu/domain/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.menu.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
