+++ src/main/java/kitchenpos/menu/dto/MenuProductRequest.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 public class MenuProductRequest {
     private Long productId;
