+++ src/main/java/kitchenpos/menu/dto/ProductRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.product.dto;
+package kitchenpos.menu.dto;
 
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 import java.math.BigDecimal;
 import java.util.Objects;
