+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -4,8 +4,8 @@ import kitchenpos.menu.domain.*;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.ProductRepository;
+import kitchenpos.menu.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
