+++ src/test/java/kitchenpos/ordering/acceptance/OrderingAcceptanceTest.java
@@ -8,9 +8,9 @@ import kitchenpos.ordering.domain.OrderLineItem;
 import kitchenpos.ordering.domain.OrderStatus;
 import kitchenpos.ordering.dto.OrderRequest;
 import kitchenpos.ordering.dto.OrderResponse;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.dto.ProductRequest;
-import kitchenpos.product.dto.ProductResponse;
+import kitchenpos.menu.domain.Product;
+import kitchenpos.menu.dto.ProductRequest;
+import kitchenpos.menu.dto.ProductResponse;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.dto.OrderTableResponse;
 import kitchenpos.utils.RestAssuredCRUD;
