+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.product.domain;
 
+import kitchenpos.menu.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
